requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
google-generativeai==0.8.4
python-dotenv==1.0.1
gunicorn==21.2.0
